requirements.txt
aiogram
python-dotenv
fastapi
uvicorn
jinja2
httpx
